+++ packages/core/src/odata/common/selectable/simple-type-fields.ts
@@ -8,6 +8,7 @@ import { DateField } from './date-field';
 import { NumberField } from './number-field';
 import { StringField } from './string-field';
 import { TimeField } from './time-field';
+import { AnyField } from './any-field';
 
 /**
  * @hidden
@@ -19,4 +20,5 @@ export type SimpleTypeFields<EntityT extends EntityBase> =
   | DateField<EntityT>
   | NumberField<EntityT>
   | StringField<EntityT>
-  | TimeField<EntityT>;
+  | TimeField<EntityT>
+  | AnyField<EntityT>;