+++ test-packages/test-services/v4/test-service/TestComplexType.d.ts
@@ -1,7 +1,7 @@
 import { Moment } from 'moment';
 import { BigNumber } from 'bignumber.js';
 import { TestNestedComplexType, TestNestedComplexTypeField } from './TestNestedComplexType';
-import { CollectionField, ComplexTypeBigNumberPropertyField, ComplexTypeBooleanPropertyField, ComplexTypeDatePropertyField, ComplexTypeField, ComplexTypeNumberPropertyField, ComplexTypeStringPropertyField, ComplexTypeTimePropertyField, ConstructorOrField, Entity, FieldType, PropertyMetadata, Time } from '@sap-cloud-sdk/core/v4';
+import { CollectionField, ComplexTypeAnyPropertyField, ComplexTypeBigNumberPropertyField, ComplexTypeBooleanPropertyField, ComplexTypeDatePropertyField, ComplexTypeField, ComplexTypeNumberPropertyField, ComplexTypeStringPropertyField, ComplexTypeTimePropertyField, ConstructorOrField, Entity, FieldType, PropertyMetadata, Time } from '@sap-cloud-sdk/core/v4';
 /**
  * TestComplexType
  */
@@ -80,6 +80,11 @@ export interface TestComplexType {
      * @nullable
      */
     sByteProperty?: number;
+    /**
+     * Geography Point Property.
+     * @nullable
+     */
+    geographyPointProperty?: any;
     /**
      * Complex Type Property.
      * @nullable
@@ -185,6 +190,11 @@ export declare class TestComplexTypeField<EntityT extends Entity> extends Comple
      * Use to reference this property in query operations such as 'filter' in the fluent request API.
      */
     sByteProperty: ComplexTypeNumberPropertyField<EntityT>;
+    /**
+     * Representation of the [[TestComplexType.geographyPointProperty]] property for query construction.
+     * Use to reference this property in query operations such as 'filter' in the fluent request API.
+     */
+    geographyPointProperty: ComplexTypeAnyPropertyField<EntityT>;
     /**
      * Representation of the [[TestComplexType.complexTypeProperty]] property for query construction.
      * Use to reference this property in query operations such as 'filter' in the fluent request API.