+++ CHANGELOG.md
@@ -18,11 +18,13 @@
 
 ## New Functionality
 
+- Include fallback `Edm.Any` for previously unsupported edm types like Edm.Geography.
+Type safety and related features like URI conversion are not supported for the fallback type.  
 - OData v4 is moven from the experimental state to GA.
 
 ## Improvements
 
--
+- Special characters like `'` (single quotes), ` ` (space) and `ä` (umlaut) are encoded properly when used in the `Filter`.
 
 ## Fixed Issues